split out mergeEdges code and tests into own module
