+++ test/unit/specs/geometry/mergeEdges.spec.js
@@ -1,87 +1,7 @@
 import _ from 'lodash';
 import { gen } from 'testcheck';
-import helpers from '../../../src/store/modules/geometry/helpers';
-import { assert, refute, nearlyEqual, assertProperty } from '../test_helpers';
-
-describe('syntheticRectangleSnaps', () => {
-  it('should work', () => {
-    const resp = helpers.syntheticRectangleSnaps(
-      /* points */ [
-        { x: 3, y: -3 }, // @
-        { x: 9, y: 11 }], // $
-      /* rectStart */ { x: 0, y: 10 },
-      /* cursorPt */ { x: 11, y: -1 },
-    );
-    /*
-              a
-           d--+---------------$
-  (0, 10) ----+---------------+---
-          |   |               |  |
-          |   |               |  |
-          |   |               |  |
-          |   |               |  |
-          |   |               |  |
-          |   |               |  |
-          |   |               |  |
-          |   |               |  |
-          |   |               |  |
-          ----+---------------+--- (11, -1)
-              |               c
-              @-------------- b
-
-    */
-    [
-      { x: 3, y: 12 }, // a
-      { x: 8, y: -3 }, // b
-      { x: 9, y: -2 }, // c
-      { x: 2, y: 11 }, // d
-    ].forEach((entry) => {
-      assert(
-        resp.find(r => (r.x === entry.x && r.y === entry.y)),
-        `expected to find ${JSON.stringify(entry)} in ${JSON.stringify(resp)}`);
-    });
-  });
-
-  it('should work with negative y vals', () => {
-    // now check with negative y vals
-    const resp = helpers.syntheticRectangleSnaps(
-      /* points */ [
-        { x: 3, y: -15 }, // @
-        { x: 9, y: -1 }], // $
-      /* rectStart */ { x: 0, y: -2 },
-      /* cursorPt */ { x: 11, y: -13 },
-    );
-
-    /*
-              a
-           d--+---------------$
-  (0, -2) ----+---------------+---
-          |   |               |  |
-          |   |               |  |
-          |   |               |  |
-          |   |               |  |
-          |   |               |  |
-          |   |               |  |
-          |   |               |  |
-          |   |               |  |
-          |   |               |  |
-          ----+---------------+--- (11, -13)
-              |               c
-              @-------------- b
-
-    */
-    [
-      { x: 3, y: 0 }, // a
-      { x: 8, y: -15 }, // b
-      { x: 9, y: -14 }, // c
-      { x: 2, y: -1 }, // d
-    ].forEach((entry) => {
-      assert(
-        resp.find(r => (r.x === entry.x && r.y === entry.y)),
-        `expected to find ${JSON.stringify(entry)} in ${JSON.stringify(resp)}`);
-    });
-  });
-});
+import { assert, refute, nearlyEqual, assertProperty } from '../../test_helpers';
+import mergeEdges from '../../../../src/store/modules/geometry/mergeEdges';
 
 describe('endpointsNearby', () => {
   const
@@ -93,7 +13,7 @@ describe('endpointsNearby', () => {
     const edge1 = { start: a, end: b };
     const edge2 = { start: c, end: d };
 
-    const res = helpers.endpointsNearby(edge1, edge2);
+    const res = mergeEdges.endpointsNearby(edge1, edge2);
     assert(
       res && res.mergeType === 'sameEndpoints',
       'expected edges to be merged');
@@ -103,14 +23,14 @@ describe('endpointsNearby', () => {
     const edge1 = { start: a, end: b };
     const edge2 = { start: d, end: c };
 
-    const res = helpers.endpointsNearby(edge1, edge2);
+    const res = mergeEdges.endpointsNearby(edge1, edge2);
     assert(
       res && res.mergeType === 'reverseEndpoints',
       'expected edges to be merged');
   });
 
   it("doesn't say endpoints are nearby when they're not", () => {
-    const res = helpers.endpointsNearby(
+    const res = mergeEdges.endpointsNearby(
       { start: a, end: c },
       { start: b, end: d });
 
@@ -118,12 +38,12 @@ describe('endpointsNearby', () => {
   });
 
   it('is sensitive to the length of an edge', () => {
-    const res = helpers.endpointsNearby(
+    const res = mergeEdges.endpointsNearby(
       { start: { x: 0, y: 0 }, end: { x: 10, y: 0 } },
       { start: { x: 0, y: 1 }, end: { x: 10, y: 1 } });
     assert(!res, "Those edges aren't nearby relative to the edge length");
 
-    const res2 = helpers.endpointsNearby(
+    const res2 = mergeEdges.endpointsNearby(
       { start: { x: 0, y: 0 }, end: { x: 100, y: 0 } },
       { start: { x: 0, y: 1 }, end: { x: 100, y: 1 } });
     assert(
@@ -134,7 +54,7 @@ describe('endpointsNearby', () => {
 
 describe('edgeDirection', () => {
   it('can handle a 45-deg angle', () => {
-    const angle = helpers.edgeDirection({
+    const angle = mergeEdges.edgeDirection({
       start: { x: 0, y: 0 },
       end: { x: 1, y: 1 },
     });
@@ -145,11 +65,11 @@ describe('edgeDirection', () => {
 
   it('can handle a straight up or straight down line', () => {
     const
-      north = helpers.edgeDirection({
+      north = mergeEdges.edgeDirection({
         start: { x: 0, y: 0 },
         end: { x: 0, y: 1 },
       }),
-      south = helpers.edgeDirection({
+      south = mergeEdges.edgeDirection({
         start: { x: 0, y: 0 },
         end: { x: 0, y: -1 },
       });
@@ -163,11 +83,11 @@ describe('edgeDirection', () => {
 
   it('can handle a west or east line', () => {
     const
-      east = helpers.edgeDirection({
+      east = mergeEdges.edgeDirection({
         start: { x: 0, y: 0 },
         end: { x: 1, y: 0 },
       }),
-      west = helpers.edgeDirection({
+      west = mergeEdges.edgeDirection({
         start: { x: 0, y: 0 },
         end: { x: -1, y: 0 },
       });
@@ -180,7 +100,7 @@ describe('edgeDirection', () => {
   });
 
   it('can handle a 30-deg angle', () => {
-    const angle = helpers.edgeDirection({
+    const angle = mergeEdges.edgeDirection({
       start: { x: 0, y: 0 },
       end: { x: Math.sqrt(3) / 2, y: 0.5 },
     });
@@ -193,13 +113,13 @@ describe('edgeDirection', () => {
 describe('haveSimilarAngles', () => {
   it('finds that identical angles are similar', () => {
     assert(
-      helpers.haveSimilarAngles(
+      mergeEdges.haveSimilarAngles(
         { start: { x: 0, y: 0 }, end: { x: 1, y: 1 } },
         { start: { x: 0, y: 0 }, end: { x: 1, y: 1 } },
       ));
 
     assert(
-      helpers.haveSimilarAngles(
+      mergeEdges.haveSimilarAngles(
         { start: { x: 0, y: 0 }, end: { x: 1, y: 5 } },
         { start: { x: -1, y: -5 }, end: { x: 0, y: 0 } },
     ));
@@ -207,7 +127,7 @@ describe('haveSimilarAngles', () => {
 
   it('recognizes that right angles are not similar', () => {
     assert(
-      !helpers.haveSimilarAngles(
+      !mergeEdges.haveSimilarAngles(
         { start: { x: 0, y: 0 }, end: { x: 1, y: 0 } },
         { start: { x: 0, y: 0 }, end: { x: 0, y: 1 } },
       ));
@@ -215,7 +135,7 @@ describe('haveSimilarAngles', () => {
 
   it('considers north and south similar', () => {
     assert(
-      helpers.haveSimilarAngles(
+      mergeEdges.haveSimilarAngles(
         { start: { x: 0, y: 0 }, end: { x: 0, y: 1 } },
         { start: { x: 0, y: 0 }, end: { x: 0, y: -1 } },
       ));
@@ -223,7 +143,7 @@ describe('haveSimilarAngles', () => {
 
   it('considers barely-north-of-east and barely-north-west similar', () => {
     assert(
-      helpers.haveSimilarAngles(
+      mergeEdges.haveSimilarAngles(
         { start: { x: 0, y: 0 }, end: { x: 100, y: 1 } },
         { start: { x: 0, y: 0 }, end: { x: -100, y: -1 } },
       ));
@@ -231,99 +151,15 @@ describe('haveSimilarAngles', () => {
 
   it('finds dissimilar angles dissimilar', () => {
     assert(
-      !helpers.haveSimilarAngles(
+      !mergeEdges.haveSimilarAngles(
         { start: { x: 0, y: 0 }, end: { x: 3, y: 2 } },
         { start: { x: 0, y: 0 }, end: { x: 2, y: 3 } },
       ));
     assert(
-      !helpers.haveSimilarAngles(
+      !mergeEdges.haveSimilarAngles(
         { start: { x: 12, y: 1 }, end: { x: 3, y: 18 } },
         { start: { x: 1, y: 24 }, end: { x: 1, y: 10 } },
       ));
-  });
-});
-
-describe('projectionOfPointToLine', () => {
-  const
-    lineUR = { p1: { x: -5, y: -5 }, p2: { x: 5, y: 5 } },
-    lineUp = { p1: { x: 0, y: -5 }, p2: { x: 0, y: 5 } },
-    lineRt = { p1: { x: -5, y: 3 }, p2: { x: 5, y: 3 } };
-  it('projects points on the line to themselves', () => {
-    const proj11 = helpers.projectionOfPointToLine({ x: 1, y: 1 }, lineUR);
-    assert(
-      nearlyEqual(proj11.x, 1) && nearlyEqual(proj11.y, 1),
-      `expected proj11 to be (1, 1) (was ${JSON.stringify(proj11)})`);
-
-    const proj33 = helpers.projectionOfPointToLine({ x: -3, y: -3 }, lineUR);
-    assert(
-      nearlyEqual(proj33.x, -3) && nearlyEqual(proj33.y, -3),
-      `expected proj33 to be (-3, -3) (was ${JSON.stringify(proj33)})`);
-  });
-
-  it('chooses the right angle projection when possible', () => {
-    const proj = helpers.projectionOfPointToLine({ x: 2, y: 0 }, lineUR);
-    assert(
-      nearlyEqual(proj.x, 1) && nearlyEqual(proj.y, 1),
-      `expected proj to be (1, 1) (was ${JSON.stringify(proj)})`);
-  });
-
-  it('projects onto verticals and horizontals', () => {
-    const genPointAboveLine = gen.object(
-      { x: gen.numberWithin(-5, 5), y: gen.int });
-    assertProperty(
-      genPointAboveLine,
-      (pt) => {
-        const projHoriz = helpers.projectionOfPointToLine(pt, lineRt);
-        assert(
-          nearlyEqual(projHoriz.x, pt.x) && nearlyEqual(projHoriz.y, 3),
-          `expected projHoriz to be (${pt.x}, 3), (was ${JSON.stringify(projHoriz)})`,
-        );
-      });
-
-    const genPointBesideLine = gen.object(
-      { x: gen.int, y: gen.numberWithin(-5, 5) });
-
-    assertProperty(
-      genPointBesideLine,
-      (pt) => {
-        const projVert = helpers.projectionOfPointToLine(pt, lineUp);
-        assert(
-          nearlyEqual(projVert.x, 0) && nearlyEqual(projVert.y, pt.y),
-          `expected projVert to be (0, ${pt.y}), (was (${JSON.stringify(projVert)})`);
-      });
-  });
-
-  it('projects points collinear with the line to an endpoint', () => {
-    const
-      collinearPtsAfterLine = gen.numberWithin(5, 10000)
-        .then(num => ({ x: num, y: num })),
-      collinearPtsBeforeLine = gen.numberWithin(-10000, -5)
-        .then(num => ({ x: num, y: num }));
-
-    assertProperty(
-      collinearPtsAfterLine,
-      (pt) => {
-        const proj = helpers.projectionOfPointToLine(pt, lineUR);
-        return nearlyEqual(proj.x, 5) && nearlyEqual(proj.y, 5);
-      });
-
-    assertProperty(
-      collinearPtsBeforeLine,
-      (pt) => {
-        const proj = helpers.projectionOfPointToLine(pt, lineUR);
-        return nearlyEqual(proj.x, -5) && nearlyEqual(proj.y, -5);
-      });
-  });
-
-  it('projects to an endpoint when necessary', () => {
-    let proj = helpers.projectionOfPointToLine({ x: 5.5, y: 5.3 }, lineUR);
-    assert(nearlyEqual(proj.x, 5) && nearlyEqual(proj.y, 5));
-
-    proj = helpers.projectionOfPointToLine({ x: 10, y: 2 }, lineUR);
-    assert(nearlyEqual(proj.x, 5) && nearlyEqual(proj.y, 5));
-
-    proj = helpers.projectionOfPointToLine({ x: 8, y: 2 }, lineRt);
-    assert(nearlyEqual(proj.x, 5) && nearlyEqual(proj.y, 3));
   });
 });
 
@@ -360,7 +196,7 @@ describe('consolidateVertices', () => {
 
   it('puts the end point in the last position', () => {
     assertProperty(genVertices, (verts) => {
-      const consolidated = helpers.consolidateVertices(...verts);
+      const consolidated = mergeEdges.consolidateVertices(...verts);
       return (
         verts[1].identity === 'end' &&
         consolidated[consolidated.length - 1].identity === 'end');
@@ -370,7 +206,7 @@ describe('consolidateVertices', () => {
   it('drops duplicate points', () => {
     assertProperty(genVertices, (verts) => {
       const
-        consolidated = helpers.consolidateVertices(...verts),
+        consolidated = mergeEdges.consolidateVertices(...verts),
         uniqVerts = _.uniqBy(verts, _.isEqual);
       assert(consolidated.length >= uniqVerts.length);
     });
@@ -378,7 +214,7 @@ describe('consolidateVertices', () => {
 
   it('never eats start and end', () => {
     assertProperty(genVertices, (verts) => {
-      const consolidated = helpers.consolidateVertices(verts[0], verts[1]);
+      const consolidated = mergeEdges.consolidateVertices(verts[0], verts[1]);
       assert(consolidated.length === 2);
     });
   });
@@ -394,7 +230,7 @@ describe('consolidateVertices', () => {
         { x: 81, y: 119, id: 'expect_gone' },
         { x: 198, y: 2, id: 'expect_gone' },
       ],
-      consolidated = helpers.consolidateVertices(...verts);
+      consolidated = mergeEdges.consolidateVertices(...verts);
 
     refute(
       _.find(consolidated, { id: 'expect_gone' }),
@@ -411,7 +247,7 @@ describe('consolidateVertices', () => {
 
 describe('oneEdgeCouldEatAnother', () => {
   it('eats edges that are on top of one another', () => {
-    const couldEat = helpers.oneEdgeCouldEatAnother(
+    const couldEat = mergeEdges.oneEdgeCouldEatAnother(
       { start: { x: 2, y: 3, id: 'a' }, end: { x: 22, y: 33, id: 'b' } },
       { start: { x: 7, y: 10.5, id: 'c' }, end: { x: 17, y: 25.5, id: 'd' } },
     );
@@ -425,7 +261,7 @@ describe('oneEdgeCouldEatAnother', () => {
   });
 
   it('wont eat edges that are far apart, even if the angle is close', () => {
-    const couldEat = helpers.oneEdgeCouldEatAnother(
+    const couldEat = mergeEdges.oneEdgeCouldEatAnother(
       { start: { x: -1000, y: 0 }, end: { x: 1000, y: 0 } },
       { start: { x: -1000, y: -100 }, end: { x: 1000, y: -100 } },
     );
@@ -433,7 +269,7 @@ describe('oneEdgeCouldEatAnother', () => {
   });
 
   it('wont eat edges with a different angle, even if the points are close', () => {
-    const couldEat = helpers.oneEdgeCouldEatAnother(
+    const couldEat = mergeEdges.oneEdgeCouldEatAnother(
       { start: { x: -1000, y: 0 }, end: { x: 1000, y: 0 } },
       { start: { x: -3, y: -1 }, end: { x: 3, y: 1 } },
     );
@@ -473,14 +309,14 @@ describe('oneEdgeCouldEatAnother', () => {
 
   it('eats edges when the small one starts near an endpoint', () => {
     assertProperty(genEdgePairSmallNearEndpoint, (edgePair) => {
-      assert(helpers.oneEdgeCouldEatAnother(...edgePair));
+      assert(mergeEdges.oneEdgeCouldEatAnother(...edgePair));
     });
   });
 });
 
 describe('edgesCombine', () => {
   it('combines edges that are parallel and nearby', () => {
-    const merge = helpers.edgesCombine(
+    const merge = mergeEdges.edgesCombine(
       { start: { x: -30, y: -1 }, end: { x: 0, y: 0 } },
       { start: { x: 30, y: 2 }, end: { x: -15, y: 0.5 } },
     );
@@ -488,7 +324,7 @@ describe('edgesCombine', () => {
   });
 
   it('combines edges that have slightly different angles and are nearby', () => {
-    const merge = helpers.edgesCombine(
+    const merge = mergeEdges.edgesCombine(
       { start: { x: -30, y: -1 }, end: { x: 0, y: 0 } },
       { start: { x: 30, y: 2 }, end: { x: -16, y: 0.5 } },
     );
@@ -497,7 +333,7 @@ describe('edgesCombine', () => {
 
 
   it("won't combine edges that have too different angles", () => {
-    const merge = helpers.edgesCombine(
+    const merge = mergeEdges.edgesCombine(
       { start: { x: -30, y: -30 }, end: { x: 0, y: 0 } },
       { start: { x: -0.3, y: -0.3 }, end: { x: 40, y: 20 } },
     );
@@ -505,7 +341,7 @@ describe('edgesCombine', () => {
   });
 
   it("won't combine edges that are too far apart", () => {
-    const merge = helpers.edgesCombine(
+    const merge = mergeEdges.edgesCombine(
       { start: { x: -5, y: 0 }, end: { x: 5, y: 10 } },
       { start: { x: 0, y: 0 }, end: { x: 10, y: 10 } },
     );
@@ -513,7 +349,7 @@ describe('edgesCombine', () => {
   });
 
   it('creates a new edge by taking the direction of most variance', () => {
-    const merge = helpers.edgesCombine(
+    const merge = mergeEdges.edgesCombine(
       { start: { x: -30, y: -1, id: 'earliest' }, end: { x: 0, y: 0 } },
       { start: { x: 30, y: 2, id: 'latest' }, end: { x: -16, y: 0.5 } },
     );
@@ -524,7 +360,7 @@ describe('edgesCombine', () => {
 
 describe('edgesExtend', () => {
   it('combines edges that almost overlap', () => {
-    const merge = helpers.edgesExtend(
+    const merge = mergeEdges.edgesExtend(
       { start: { x: -30, y: -1 }, end: { x: 0, y: 0 } },
       { start: { x: 0.5, y: 0.5 }, end: { x: 31, y: 1 } },
     );
@@ -532,7 +368,7 @@ describe('edgesExtend', () => {
   });
 
   it('combines edges that almost overlap (even when one needs to reverse)', () => {
-    const merge = helpers.edgesExtend(
+    const merge = mergeEdges.edgesExtend(
       { start: { x: -30, y: -1 }, end: { x: 0, y: 0 } },
       { start: { x: 31, y: 1 }, end: { x: 0.5, y: 0.5 } },
     );
@@ -540,7 +376,7 @@ describe('edgesExtend', () => {
   });
 
   it('ignores edges that almost overlap, but have different angles', () => {
-    const merge = helpers.edgesExtend(
+    const merge = mergeEdges.edgesExtend(
       { start: { x: -30, y: -1 }, end: { x: 0, y: 0 } },
       { start: { x: 2, y: 31 }, end: { x: 0.5, y: 0.5 } },
     );
@@ -548,7 +384,7 @@ describe('edgesExtend', () => {
   });
 
   it('ignores edges that do not nearly-overlap', () => {
-    const merge = helpers.edgesExtend(
+    const merge = mergeEdges.edgesExtend(
       { start: { x: 0, y: 0 }, end: { x: 3, y: 0 } },
       { start: { x: 4, y: 0 }, end: { x: 5, y: 0 } },
     );
